requirements.txt
fastapi>=0.95.0,<0.96.0
uvicorn>=0.22.0,<0.23.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.5,<3.0.0
pgvector>=0.2.0,<0.3.0
python-multipart>=0.0.6,<0.1.0
python-dotenv>=1.0.0,<2.0.0
boto3>=1.28.0,<2.0.0
requests>=2.31.0,<3.0.0
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<0.16.0
ftfy>=6.1.0,<7.0.0
regex>=2023.0.0
tqdm>=4.65.0,<5.0.0
pytest>=7.3.1,<8.0.0
httpx>=0.24.0,<0.25.0
pytest-asyncio>=0.21.0,<0.22.0
git+https://github.com/openai/CLIP.git
